requirements-build.txt
pyserial
pyinstaller
pillow
